+++ jumpscale/packages/tfgrid_solutions/chats/gitea_advanced.py
@@ -45,35 +45,11 @@ class Gitea(MarketPlaceAppsChatflow):
         self.database_user = self.database_user.value
         self.database_password = self.database_password.value
         self.repository_name = self.repository_name.value
-
-    @chatflow_step(title="SSH key (Optional)")
-    def upload_public_key(self):
-        self.public_key = (
-            self.upload_file(
-                "Please upload your public ssh key, this will allow you to access your threebot container using ssh",
-            )
-            or ""
-        )
-        self.public_key = self.public_key.strip()
+        self.user_email = self.user_info()["email"]
 
     @chatflow_step(title="New Expiration")
     def set_expiration(self):
         self.expiration = deployer.ask_expiration(self)
-
-    @chatflow_step(title="Backup credentials")
-    def backup_credentials(self):
-        form = self.new_form()
-        aws_access_key_id = form.string_ask("AWS access key id", required=True)
-        aws_secret_access_key = form.secret_ask("AWS secret access key", required=True)
-        restic_password = form.secret_ask("Restic Password", required=True)
-        restic_repository = form.string_ask(
-            "Restic Repository Example: `s3:s3backup.tfgw-testnet-01.gateway.tf/testbucket`", required=True, md=True
-        )  # TODO: AUTOMATE THAT
-        form.ask("These credentials will be used to backup your solution.", md=True)
-        self.aws_access_key_id = aws_access_key_id.value
-        self.aws_secret_access_key = aws_secret_access_key.value
-        self.restic_password = restic_password.value
-        self.restic_repository = restic_repository.value
 
     @chatflow_step(title="Initializing backup")
     def init_backup(self):
@@ -174,7 +150,7 @@ class Gitea(MarketPlaceAppsChatflow):
             network_name=self.network_view.name,
             trc_secret=self.secret,
             domain=self.domain,
-            email=self.user_info()["email"],
+            email=self.user_email,
             solution_ip=self.ip_address,
             solution_port=3000,
             enforce_https=True,
@@ -192,29 +168,6 @@ class Gitea(MarketPlaceAppsChatflow):
                 solution_uuid=self.solution_id,
                 wid=self.reverse_proxy_id,
             )
-
-    @chatflow_step(title="Initializing backup")
-    def init_backup(self):
-        solution_name = self.solution_name.replace(".", "_").replace("-", "_")
-        self.md_show_update("Setting container backup")
-        SOLUTIONS_WATCHDOG_PATHS = j.sals.fs.join_paths(j.core.dirs.VARDIR, "solutions_watchdog")
-        if not j.sals.fs.exists(SOLUTIONS_WATCHDOG_PATHS):
-            j.sals.fs.mkdirs(SOLUTIONS_WATCHDOG_PATHS)
-
-        restic_instance = j.tools.restic.get(solution_name)
-        restic_instance.password = self.restic_password
-        restic_instance.repo = self.restic_repository
-        restic_instance.extra_env = {
-            "AWS_ACCESS_KEY_ID": self.aws_access_key_id,
-            "AWS_SECRET_ACCESS_KEY": self.aws_secret_access_key,
-        }
-        restic_instance.save()
-        try:
-            restic_instance.init_repo()
-        except Exception as e:
-            j.tools.restic.delete(solution_name)
-            raise j.exceptions.Input(f"Error: Failed to reach repo {self.restic_repository} due to {str(e)}")
-        restic_instance.start_watch_backup(SOLUTIONS_WATCHDOG_PATHS)
 
     @chatflow_step(title="Success", disable_previous=True, final_step=True)
     def success(self):
